requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
attrs==20.1.0
awscli==1.18.100
backcall==0.2.0
bleach==3.1.5
botocore==1.17.23
cffi==1.14.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-filter==2.3.0
djangorestframework==3.11.1
djongo==1.3.3
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
importlib-metadata==1.6.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.6
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
pymongo==3.11.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
rsa==4.5
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.2.4
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
virtualenv==20.0.23
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
